requirements.txt
Pillow==10.2.0
pytesseract==0.3.10
python-docx==1.1.0 